Split uigf api and json functions out to remove circular dependency
